+++ container_transform/tests/converter_tests.py
@@ -1,3 +1,4 @@
+import json
 from unittest import TestCase
 
 from container_transform.converter import Converter
@@ -42,3 +43,28 @@ class ConverterTests(TestCase):
         output = conv.convert()
 
         self.assertIsInstance(output, str)
+
+    def test_compose_converter_v2_to_ecs(self):
+        self.maxDiff = None
+
+        filename = './container_transform/tests/composev2_extended.yml'
+        output_filename = './container_transform/tests/composev2_extended_output.json'
+        conv = Converter(filename, 'compose', 'ecs')
+
+        output = conv.convert()
+        output_dict = json.loads(output)
+
+        output_want = json.load(open(output_filename, 'r'))
+        self.assertDictEqual(output_dict, output_want)
+
+    def test_compose_converter_v2_systemd(self):
+        self.maxDiff = None
+
+        filename = './container_transform/tests/composev2.yml'
+        output_filename = './container_transform/tests/composev2_output.service'
+        conv = Converter(filename, 'compose', 'systemd')
+
+        output = conv.convert()
+
+        output_want = open(output_filename, 'r').read()
+        self.assertEqual(output, output_want)